\name{dtwar-package}
\alias{dtwar-package}
\alias{dtwar}
\docType{package}
\title{
\packageTitle{dtwar}
}
\description{
\packageDescription{dtwar}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dtwar}
\packageIndices{dtwar}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{dtwar}

Maintainer: \packageMaintainer{dtwar}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
